Restore original autosteal mixin
