+++ src/main/resources/wurst.mixins.json
@@ -21,6 +21,7 @@
     "ClientPlayerEntityMixin",
     "ClientPlayerInteractionManagerMixin",
     "ClientPlayNetworkHandlerMixin",
+    "ContainerScreen54Mixin",
     "CustomPayloadC2SPacketAccessor",
     "DeathScreenMixin",
     "DirectConnectScreenMixin",
@@ -32,7 +33,6 @@
     "FluidRendererMixin",
     "GameMenuScreenMixin",
     "GameRendererMixin",
-    "HandledScreenMixin",
     "IngameHudMixin",
     "InGameOverlayRendererMixin",
     "KeyBindingMixin",
